#1992 - Fix regular bows being used in water and on other classes
